requirements.txt
certifi==2019.11.28
cffi==1.14.0
cloudpickle==1.2.2
future==0.18.2
gym==0.16.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
pycparser==2.19
pyglet==1.4.10
PyYAML==5.3
scipy==1.4.1
selenium==3.141.0
six==1.14.0
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.8
wincertstore==0.2
